2回目提出

+++ 社内研修(Java)/Stage3-C/メソッドへのふるまい追加.java
@@ -3,16 +3,22 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.Collections;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 
-class Main {
+class Stage3C {
     public static void main (String[] args) {
         printColors("red","blue","yellow","green");
     }
     
     private static void printColors(String ... colors) {
-    　　//テキストファイルを作成
+        //現在の日付を取得
+        Date date = new Date();
+        //日付フォーマットを指定
+        SimpleDateFormat sdf = new SimpleDateFormat("yyMMdd");
+        //テキストファイルを作成
         try {
-        File file = new File("/home/ec2-user/environment/サンプル実行/output/{200427}_output.txt");
+        File file = new File("/home/ec2-user/environment/サンプル実行/output/{" + sdf.format(date) + "}_output.txt");
         FileWriter fw = new FileWriter(file);
         for (int i = 0; i < colors.length; i++) {
              fw.write(colors[i]);
@@ -26,6 +32,8 @@ class Main {
         File dir = new File("/home/ec2-user/environment/サンプル実行/output");
         File[] f = dir.listFiles();
         Arrays.sort(f, Collections.reverseOrder());
+        if(f.length >= 5){
         f[4].delete();
+        }
     }
 }
